requirements.txt
matplotlib
norse
tonic
numpy
pandas
pysdl2
pysdl2-dll
ipykernel